+++ src/crh/bowe_hopwood/constraints.rs
@@ -46,7 +46,7 @@ where
     P: TEModelParameters,
     W: Window,
 {
-    type OutputVar = AffineVar<P, F>;
+    type OutputVar = F;
     type ParametersVar = ParametersVar<P, W>;
 
     #[tracing::instrument(target = "r1cs", skip(parameters, input))]
@@ -81,7 +81,7 @@ where
             &input_in_bits,
         )?;
 
-        Ok(result)
+        Ok(result.x)
     }
 }
 
@@ -111,7 +111,6 @@ mod test {
     use crate::crh::bowe_hopwood;
     use crate::crh::pedersen;
     use crate::{CRHGadget, CRH};
-    use ark_ec::ProjectiveCurve;
     use ark_ed_on_bls12_381::{constraints::FqVar, EdwardsParameters, Fq as Fr};
     use ark_r1cs_std::{alloc::AllocVar, uint8::UInt8, R1CSVar};
     use ark_relations::r1cs::{ConstraintSystem, ConstraintSystemRef};
@@ -167,8 +166,7 @@ mod test {
 
         println!("number of constraints total: {}", cs.num_constraints());
 
-        let primitive_result = primitive_result.into_affine();
-        assert_eq!(primitive_result, result_var.value().unwrap().into_affine());
+        assert_eq!(primitive_result, result_var.value().unwrap());
         assert!(cs.is_satisfied().unwrap());
     }
 }