+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Breaking changes
 
+- [\#56](https://github.com/arkworks-rs/crypto-primitives/pull/56) Compress the output of the Bowe-Hopwood-Pedersen CRH to a single field element, in line with the Zcash specification.
+
 ### Features
 
 ### Improvements